qos: Disable irq notifier when qos request is removed
When a qos request is added with IRQ affinity, irq notifier for that irq
is added and without disabling the irq notifier if the irq is free it
throws a warning. So disable irq notifier when the qos request is
removed.

Change-Id: I50faa4ecbe1b632c0f0f203ca52faf18753c33d4
Signed-off-by: Srinivas Rao L <lsrao@codeaurora.org>
Signed-off-by: Lina Iyer <ilina@codeaurora.org>
[fix checkpatch issues]
[Sultan Alsawaf: forward port to 5.10]
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
(cherry picked from commit 7f5e2b50dd8bd7c04d7ad569beebd1e051513f1e)