vh: sched: Support vendor group util
Cpu cgropus provide per-group util, however, it will combine with
cpu.shares. If cpu.shares is not enabled, then we could not use
per-group util. Besides, it does not provide the estimated util
per group. This patch implements the vendor per-group util as well
as per-group estimated util. With this, we could control the max
util per group for power consideration.

Also to reduce the overheadh, we simply the groups into 2 virtual
groups only.

Bug: 201260585
Change-Id: I3a1ce1af902d59473dcf6e6da1a630c240f09dfb
Signed-off-by: Rick Yiu <rickyiu@google.com>
Signed-off-by: acuicultor <ramoncuenca4@gmail.com>

# Conflicts:
#	drivers/soc/google/vh/kernel/sched/fair.c
#	drivers/soc/google/vh/kernel/sched/init.c