kbuild: Implement support for KBUILD_OPTIONS
GKI modules are all compiled out-of-tree and define their own local set of
CONFIGs which are passed to the make command itself when the module is
built. KBUILD_OPTIONS is consistently used as a local Makefile variable to
store all of the module's locally configured CONFIGs to pass to the make
command line.

GKI modules thus do not use the Kconfig system to define their CONFIGs, and
instead rely exclusively on the KBUILD_OPTIONS idiom, which makes it hard
to compile these modules in-tree. Some modules also configure CONFIGs in
KBUILD_OPTIONS which are not meant to be defined tree-wide, so even if all
the necessary CONFIGs are properly defined via Kconfig, they would have
tree-wide effects which are not desired.

To make it possible to compile GKI modules in-tree, implement support for
the KBUILD_OPTIONS variable directly into the build system. Makefiles which
define KBUILD_OPTIONS will have all variables in KBUILD_OPTIONS expanded
and passed via the make command line in an additional pass, so that all the
CONFIGs in KBUILD_OPTIONS will be defined for the Makefile itself which
defines KBUILD_OPTIONS.

Furthermore, when there's a Kbuild in the same directory as a Makefile that
defines KBUILD_OPTIONS, the Makefile will take precedence over the Kbuild
for the first make pass, and will still be included in addition to the
Kbuild for the second make pass. This differs from normal Kbuild behavior
where Kbuild always takes precedence over Makefile in the same directory.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
(cherry picked from commit 5f56f250f496714ef8c20030f1103b25da5501af)